+++ cli/CHANGELOG.md
@@ -2,7 +2,15 @@
 
 This change log covers only the command line interface (CLI) of Open VSX.
 
-### [next] (unreleased)
+### [v1.1.1] (09/08/2026)
+
+### Dependencies
+
+- Bump ip-address from `10.2.0` to `10.4.0`
+- Bump fast-uri from `3.1.4` to `3.1.5`
+- Bump js-yaml from `4.3.0` to `4.3.1`
+
+### [v1.1.0] (02/08/2026)
 
 #### Added
 
@@ -19,6 +27,7 @@ This change log covers only the command line interface (CLI) of Open VSX.
 
 - Bump js-yaml from `4.2.0` to `4.3.0` ([#1976](https://github.com/eclipse/openvsx/pull/1976))
 - Bump tar from `7.5.16` to `7.5.21` ([#1987](https://github.com/eclipse/openvsx/pull/1987))
+- Bump brace-expansion from `1.1.16` to `1.1.18` ([#2031](https://github.com/eclipse/openvsx/pull/2031))
 
 ### [v1.0.2] (23/06/2026)